quiz_backend/pyproject.toml
[tool.poetry]
name = "quiz-backend"
version = "0.1.0"
description = ""
authors = ["Hammad-Rehman <124163466+developer-hammad-rehman@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
sqlmodel = "^0.0.16"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-jose = "^3.3.0"
psycopg2 = "^2.9.9"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
